Move vcs manager component to view module
