+++ src/browser/vcs/_vcs-manager-sizes.scss
