+++ src/browser/vcs/index.ts
@@ -5,3 +5,4 @@ export * from './vcs.state';
 export * from './vcs.actions';
 export * from './vcs.reducer';
 export * from './vcs.effects';
+export * from './vcs-manager.component';
