+++ src/browser/vcs/vcs-manager.component.html
