+++ src/browser/vcs/vcs-manager.component.scss
@@ -1,5 +1,5 @@
-@import "../../../ui/style/spacing";
-@import "../../../ui/style/typography";
+@import "../ui/style/spacing";
+@import "../ui/style/typography";
 @import "./vcs-manager-sizes";
 
 .VcsManager {
