+++ src/browser/vcs/vcs-view/index.ts
@@ -1,6 +1,6 @@
 export * from './vcs-view.module';
-export * from './vcs-manager/vcs-manager.component';
 export * from './base-vcs-item/base-vcs-item-factory';
 export * from './base-vcs-item/base-vcs-item.component';
 export * from './vcs-item-maker';
 export * from './vcs-item';
+export * from './vcs-item-list-manager';
