+++ src/browser/vcs/vcs-view/vcs-view.module.ts
@@ -4,7 +4,6 @@ import { BaseVcsItemFactory } from './base-vcs-item/base-vcs-item-factory';
 import { BaseVcsItemComponent } from './base-vcs-item/base-vcs-item.component';
 import { VcsItemListManager } from './vcs-item-list-manager';
 import { VcsItemMaker } from './vcs-item-maker';
-import { VcsManagerComponent } from './vcs-manager/vcs-manager.component';
 
 
 @NgModule({
@@ -12,11 +11,9 @@ import { VcsManagerComponent } from './vcs-manager/vcs-manager.component';
         UiModule,
     ],
     declarations: [
-        VcsManagerComponent,
         BaseVcsItemComponent,
     ],
     entryComponents: [
-        VcsManagerComponent,
         BaseVcsItemComponent,
     ],
     providers: [
@@ -25,7 +22,6 @@ import { VcsManagerComponent } from './vcs-manager/vcs-manager.component';
         VcsItemListManager,
     ],
     exports: [
-        VcsManagerComponent,
         BaseVcsItemComponent,
     ],
 })
