+++ src/browser/vcs/vcs.module.ts
@@ -1,7 +1,9 @@
 import { NgModule } from '@angular/core';
 import { EffectsModule } from '@ngrx/effects';
 import { StoreModule } from '@ngrx/store';
+import { UiModule } from '../ui/ui.module';
 import { VcsAuthenticationDatabaseProvider } from './vcs-authentication-database';
+import { VcsManagerComponent } from './vcs-manager.component';
 import { VcsRemoteModule } from './vcs-remote';
 import { VcsViewModule } from './vcs-view';
 import { VcsEffects } from './vcs.effects';
@@ -11,10 +13,17 @@ import { VcsService } from './vcs.service';
 
 @NgModule({
     imports: [
+        UiModule,
         VcsRemoteModule,
         VcsViewModule,
         StoreModule.forFeature('vcs', vcsReducerMap),
         EffectsModule.forFeature([VcsEffects]),
+    ],
+    declarations: [
+        VcsManagerComponent,
+    ],
+    entryComponents: [
+        VcsManagerComponent,
     ],
     providers: [
         VcsAuthenticationDatabaseProvider,
@@ -23,6 +32,7 @@ import { VcsService } from './vcs.service';
     exports: [
         VcsRemoteModule,
         VcsViewModule,
+        VcsManagerComponent,
     ],
 })
 export class VcsModule {
